requirements.txt
fastapi==0.114.2
httpx==0.27.2
pydantic==2.9.1
pydantic_settings==2.5.2
pytest==8.3.2
python-dotenv==1.0.1
uvicorn==0.30.6
